expense-tracker-bot/requirements.txt
python-telegram-bot[job-queue]
huggingface-hub
requests
pandas
python-dotenv
Pillow